Move Widget / CheckboxInput / RadioSelect / Select tests into components/Widget and decouple from telepath
